requirements.txt
pandas==2.2.1
psycopg2-binary==2.9.9
python-pptx==0.6.23
